config: enable DVB/Media devices
We seem to not have any significant amount of DVB and media devices
enabled in our main bare metal kernel images, which means that many
video capture cards do not work.

I've gone through the selection of the kernel for the most relevant
cards that I can find (of which I have some) and verified that all the
relevant options related to them are enabled, and diffed out the
options. It's a long list, but there isn't much to this than modules, only.

Undoubtably there's a few cards/devices in the list that are on
platforms that we don't support, but that doesn't really hurt much. Most
of the really popular TV/FM tuner and video capture cards are on here.

Thanks,

Auke